requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
matplotlib>=3.8.0
python-pptx>=0.6.23
apscheduler>=3.10.0
